修改查询期号和算奖BUG

+++ src/Baibaocp.Storaging/Entities/Merchants/Merchanter.cs
@@ -47,18 +47,18 @@ namespace Baibaocp.Storaging.Entities.Merchants
         /// 预存款余额
         /// </summary>
         [Column("RestPreMoney")]
-        public int Balance { get; set; }
+        public decimal Balance { get; set; }
 
         /// <summary>
         /// 总出票金额
         /// </summary>
         [Column("OutTicketMoney")]
-        public int TotalTicketedAmount { get; set; }
+        public decimal TotalTicketedAmount { get; set; }
 
         /// <summary>
         /// 总返奖金额
         /// </summary>
         [Column("RewardMoney")]
-        public int TotalAwardedAmount { get; set; }
+        public decimal TotalAwardedAmount { get; set; }
     }
 }
